[Dev] Move to Storybooks 3.2
