+++ package.json
@@ -99,8 +99,8 @@
     "webpack-stream": "^3.2.0"
   },
   "dependencies": {
-    "@storybook/addon-options": "^3.1.2",
-    "@storybook/react": "^3.1.3",
+    "@storybook/addon-options": "^3.2.0",
+    "@storybook/react": "^3.2.0",
     "artsy-passport": "^2.0.3",
     "artsy-xapp": "^1.0.4",
     "babel-polyfill": "^6.23.0",
