+++ src/components/__stories__/artwork.story.tsx
