+++ src/components/__stories__/artwork_grid.story.tsx
