+++ src/components/__stories__/fair_booth.story.tsx
