+++ src/components/__stories__/gene.story.tsx
