+++ src/components/__stories__/inquiry_artwork.story.tsx
