+++ src/components/publishing/__test__/typography.test.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import * as renderer from "react-test-renderer"
 
-import Typography from "../__stories__/typography"
+import Typography from "../__stories__/typography.story"
 
 it("renders properly", () => {
   const typography = renderer.create(<Typography />).toJSON()
